+++ tests/test_uncertainties.py
@@ -1,6 +1,5 @@
 import copy
 import json
-import inspect
 import math
 from pathlib import Path
 import random  # noqa
@@ -12,7 +11,6 @@ from uncertainties.core import (
     ufloat,
     AffineScalarFunc,
     ufloat_fromstr,
-    deprecated_methods,
 )
 from uncertainties import (
     umath,
@@ -338,44 +336,16 @@ def test_pickling():
     assert pickle.loads(pickle.dumps(x)).linear_combo == {}
 
 
-def test_int_div():
-    "Integer division"
-    # We perform all operations on floats, because derivatives can
-    # otherwise be meaningless:
-    x = ufloat(3.9, 2) // 2
-    assert x.nominal_value == 1.0
-    # All errors are supposed to be small, so the ufloat()
-    # in x violates the assumption.  Therefore, the following is
-    # correct:
-    assert x.std_dev == 0.0
-
-
 def test_comparison_ops():
     "Test of comparison operators"
 
     # Operations on quantities equivalent to Python numbers must still
     # be correct:
-    a = ufloat(-3, 0)
     b = ufloat(10, 0)
     c = ufloat(10, 0)
-    assert a < b
-    assert a < 3
-    assert 3 < b  # This is first given to int.__lt__()
     assert b == c
 
     x = ufloat(3, 0.1)
-
-    # One constraint is that usual Python code for inequality testing
-    # still work in a reasonable way (for instance, it is generally
-    # desirable that functions defined by different formulas on
-    # different intervals can still do "if 0 < x < 1:...".  This
-    # supposes again that errors are "small" (as for the estimate of
-    # the standard error).
-    assert x > 1
-
-    # The limit case is not obvious:
-    assert not (x >= 3)
-    assert not (x < 3)
 
     assert x == x
     # Comparaison between Variable and AffineScalarFunc:
@@ -427,7 +397,7 @@ def test_comparison_ops():
             return (random.random() - 0.5) * min(var.std_dev, 1e-5) + var.nominal_value
 
         # All operations are tested:
-        for op in ["__%s__" % name for name in ("ne", "eq", "lt", "le", "gt", "ge")]:
+        for op in ["__%s__" % name for name in ("ne", "eq")]:
             try:
                 float_func = getattr(float, op)
             except AttributeError:  # Python 2.3's floats don't have __ne__
@@ -484,17 +454,16 @@ def test_comparison_ops():
 
 
 def test_logic():
-    "Boolean logic: __nonzero__, bool."
-
+    "bool defers to object.__bool__ and always returns True."
     x = ufloat(3, 0)
     y = ufloat(0, 0)
     z = ufloat(0, 0.1)
     t = ufloat(-1, 2)
 
     assert bool(x)
-    assert not bool(y)
+    assert bool(y)
     assert bool(z)
-    assert bool(t)  # Only infinitseimal neighborhood are used
+    assert bool(t)
 
 
 def test_basic_access_to_data():
@@ -1116,20 +1085,6 @@ else:
         assert len(numpy.array([x, x, x]) == x) == 3
         assert numpy.all(x == numpy.array([x, x, x]))
 
-        # Inequalities:
-        assert len(x < numpy.arange(10)) == 10
-        assert len(numpy.arange(10) > x) == 10
-        assert len(x <= numpy.arange(10)) == 10
-        assert len(numpy.arange(10) >= x) == 10
-        assert len(x > numpy.arange(10)) == 10
-        assert len(numpy.arange(10) < x) == 10
-        assert len(x >= numpy.arange(10)) == 10
-        assert len(numpy.arange(10) <= x) == 10
-
-        # More detailed test, that shows that the comparisons are
-        # meaningful (x >= 0, but not x <= 1):
-        assert numpy.all((x >= numpy.arange(3)) == [True, False, False])
-
     def test_correlated_values():
         """
         Correlated variables.
@@ -1344,18 +1299,6 @@ def test_no_numpy():
         match="not able to import numpy",
     ):
         _ = correlation_matrix([x, y, z])
-
-
-@pytest.mark.parametrize("method_name", deprecated_methods)
-def test_deprecated_method(method_name):
-    x = ufloat(1, 0.1)
-    y = ufloat(-12, 2.4)
-    num_args = len(inspect.signature(getattr(float, method_name)).parameters)
-    with pytest.warns(FutureWarning, match="will be removed"):
-        if num_args == 1:
-            getattr(x, method_name)()
-        else:
-            getattr(x, method_name)(y)
 
 
 def test_zero_std_dev_warn():